Feat: 시간표 목표 설정

+++ app/src/main/java/com/example/relay/timetable/models/MemSchedule.kt
@@ -2,9 +2,9 @@ package com.example.relay.timetable.models
 
 data class MemSchedule(
     val timeTableIdx:Long,
-    val day:Long,
+    val day:Int,
     val start:String,
     val end:String,
-    val goal:Long,
+    val goal:Float,
     val goalType:String
 )
